tidying plus comments
